+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/life/PoadController.java
@@ -1,0 +1,143 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.life;
+
+import com.dwarfeng.familyhelper.life.sdk.bean.entity.JSFixedFastJsonPoad;
+import com.dwarfeng.familyhelper.life.stack.bean.entity.Poad;
+import com.dwarfeng.familyhelper.life.stack.bean.key.PoadKey;
+import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.life.JSFixedFastJsonDispPoad;
+import com.dwarfeng.familyhelper.webapi.stack.bean.disp.life.DispPoad;
+import com.dwarfeng.familyhelper.webapi.stack.handler.system.TokenHandler;
+import com.dwarfeng.familyhelper.webapi.stack.service.life.PoadResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 活动数据权限控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.10
+ */
+@RestController
+@RequestMapping("/api/v1/life")
+public class PoadController {
+
+    private final PoadResponseService service;
+
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<Poad, JSFixedFastJsonPoad> poadBeanTransformer;
+    private final BeanTransformer<DispPoad, JSFixedFastJsonDispPoad> dispPoadBeanTransformer;
+
+    private final TokenHandler tokenHandler;
+
+    public PoadController(
+            PoadResponseService service,
+            ServiceExceptionMapper sem,
+            BeanTransformer<Poad, JSFixedFastJsonPoad> poadBeanTransformer,
+            BeanTransformer<DispPoad, JSFixedFastJsonDispPoad> dispPoadBeanTransformer,
+            TokenHandler tokenHandler
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.poadBeanTransformer = poadBeanTransformer;
+        this.dispPoadBeanTransformer = dispPoadBeanTransformer;
+        this.tokenHandler = tokenHandler;
+    }
+
+    @GetMapping("/poad/{longId}&{stringId}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            boolean exists = service.exists(new PoadKey(longId, stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/poad/{longId}&{stringId}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPoad> get(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            Poad poad = service.get(new PoadKey(longId, stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPoad.of(poad)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/poad/{longId}&{stringId}/disp")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonDispPoad> getDisp(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            StringIdKey inspectAccountKey = tokenHandler.getAccountKey(request);
+            DispPoad dispPoad = service.getDisp(new PoadKey(longId, stringId), inspectAccountKey);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonDispPoad.of(dispPoad)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-data-set/{activityDataSetId}/poad")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonPoad>> childForActivityDataSet(
+            HttpServletRequest request, @PathVariable("activityDataSetId") Long activityDataSetId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<Poad> childForActivityDataSet = service.childForActivityDataSet(
+                    new LongIdKey(activityDataSetId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonPoad> transform = PagingUtil.transform(
+                    childForActivityDataSet, poadBeanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-data-set/{activityDataSetId}/poad/disp")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonDispPoad>> childForActivityDataSetDisp(
+            HttpServletRequest request, @PathVariable("activityDataSetId") Long activityDataSetId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            StringIdKey inspectAccountKey = tokenHandler.getAccountKey(request);
+            PagedData<DispPoad> childForActivityDataSetDisp = service.childForActivityDataSetDisp(
+                    new LongIdKey(activityDataSetId), new PagingInfo(page, rows), inspectAccountKey
+            );
+            PagedData<JSFixedFastJsonDispPoad> transform = PagingUtil.transform(
+                    childForActivityDataSetDisp, dispPoadBeanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}