+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - 新建 Controller。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDriverSupportController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTypeIndicatorController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityDataSetController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PbSetController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoadController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。